+++ src/CloudService/exports/Get-AzCloudServiceNetworkInterfaces.ps1
@@ -78,7 +78,7 @@ CLOUDSERVICE <CloudService>: CloudService instance.
     [(Any) <String>]: This indicates any property can be added to this object.
   [UpgradeMode <CloudServiceUpgradeMode?>]: Update mode for the cloud service. Role instances are allocated to update domains when the service is deployed. Updates can be initiated manually in each update domain or initiated automatically in all update domains.         Possible Values are <br /><br />**Auto**<br /><br />**Manual** <br /><br />**Simultaneous**<br /><br />         If not specified, the default value is Auto. If set to Manual, PUT UpdateDomain must be called to apply the update. If set to Auto, the update is automatically applied to each update domain in sequence.
 .Link
-https://docs.microsoft.com/en-us/powershell/module/az.cloudservice/get-AzCloudServiceNetworkInterfaces
+https://docs.microsoft.com/powershell/module/az.cloudservice/get-AzCloudServiceNetworkInterfaces
 #>
 function Get-AzCloudServiceNetworkInterfaces {
 [CmdletBinding(DefaultParameterSetName='CloudServiceName', PositionalBinding=$false)]