+++ src/Compute/Compute.Test/ScenarioTests/VirtualMachineTests.cs
@@ -367,6 +367,13 @@ namespace Microsoft.Azure.Commands.Compute.Test.ScenarioTests
         {
             TestRunner.RunTestScript("Test-VirtualMachineGetVMExtensionPiping");
         }
+        
+        [Fact]
+        [Trait(Category.AcceptanceType, Category.CheckIn)]
+        public void TestVirtualMachinePatchAPI()
+        {
+            TestRunner.RunTestScript("Test-VirtualMachinePatchAPI");
+        }
 
         [Fact]
         [Trait(Category.AcceptanceType, Category.CheckIn)]