+++ src/MySql/exports/Test-AzMySqlFlexibleServerConnect.ps1
@@ -65,7 +65,7 @@ INPUTOBJECT <IMySqlIdentity>: The server to connect.
   [SubscriptionId <String>]: The ID of the target subscription.
   [VirtualNetworkRuleName <String>]: The name of the virtual network rule.
 .Link
-https://docs.microsoft.com/en-us/powershell/module/az.mysql/test-azmysqlflexibleserverconnect
+https://docs.microsoft.com/powershell/module/az.mysql/test-azmysqlflexibleserverconnect
 #>
 function Test-AzMySqlFlexibleServerConnect {
 [OutputType([System.String])]