+++ src/Blueprint/Blueprint/ChangeLog.md
@@ -18,6 +18,8 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+
+## Version 0.3.0
 * Added a fix to unregister the API delegating handler from blueprint client
 
 ## Version 0.2.13