+++ src/Storage/Storage.Management/ChangeLog.md
@@ -18,6 +18,10 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Fixed an issue that list account from resource group won't use nextlink
+    - `Get-AzStorageAccount`
+
+## Version 3.4.0
 * Upgraded to Microsoft.Azure.Management.Storage 19.0.0, to support new API version 2021-01-01.
 * Supported resource access rule in NetworkRuleSet
     - `Update-AzStorageAccountNetworkRuleSet`